Choose the correct answer challenges in English and Spanish
